feat: CanvasSection 컴포넌트 activeTool, 사이드이펙트 분리 #51
- CanvasSection 컴포넌트 activeTool, sideEffect를 Toolbar 컴포넌트로 분리
